longitude latidute distance calc

+++ src/main/java/airlineManager/AirportsLoader.java
@@ -29,7 +29,7 @@ public class AirportsLoader implements InterfaceGameFileLoader<Airport> {
                 airports.add(
                     new Airport(airportInfo[3], 
                                 Integer.parseInt(airportInfo[4]), Integer.parseInt(airportInfo[5]),
-                                airportInfo[0], Integer.parseInt(airportInfo[1]), Integer.parseInt(airportInfo[2])
+                                airportInfo[0], Float.parseFloat(airportInfo[1]), Float.parseFloat(airportInfo[2])
                     )
                 );
